homework-api/homework-api/pyproject.toml
[tool.poetry]
name = "fast-api-boilerplate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
#psycopg2-binary = "^2.9.2"
python-dotenv = "^0.19.2"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
